Everything works yeeeeeeeeeeeaaaaaaaaaaaaaaaaahhhhhhhhhhhhhhhhhhhhhhhhhhhhhhhhhhhhhhhhhhhhhhhhhhhhhhhhhhhhhhhhhhhhhhhhhhhhhhhhhhhhhhhhhhhhhhhhhhhhhhhhhhhhhhhhhhhhhhhhhh me happy now going for a night sleep (at 4 O'clock in the morning) O_O

+++ MatchMusicTaste_v8/src/main/java/nl/han/ica/daretwodate/services/ITasteOMeter.java
@@ -6,6 +6,6 @@ import java.util.List;
  */
 public interface ITasteOMeter {
     boolean callTasteOMeter(String user1, String user2, String limit, String lastfmApiKey);
-    float getTasteOMeterPercentage();
+    int getTasteOMeterPercentage();
     List<String> getTasteOMeterArtists();
 }
